Daniel dev

+++ multistock-backend/database/migrations/2025_05_22_211640_create_product_sale_table.php
@@ -11,15 +11,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('product_sale', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('product_id')->constrained('stock_warehouses')->onDelete('cascade');
-            $table->foreignId('venta_id')->constrained('sale')->onDelete('cascade');
-            $table->integer('cantidad');
-            $table->decimal('precio_unidad', 10, 2);
-            $table->decimal('precio_total',10,2);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('product_sale')) {
+            Schema::create('product_sale', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('product_id')->constrained('stock_warehouses')->onDelete('cascade');
+                $table->foreignId('venta_id')->constrained('sale')->onDelete('cascade');
+                $table->integer('cantidad');
+                $table->decimal('precio_unidad', 10, 2);
+                $table->decimal('precio_total',10,2);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
